base: Allow QS tile reordering
This commit makes tiles resond to a long-press. That starts a drag
operation on them. Dropping the tile places it somewhere. The anims
are fairly simple - a nice fade for the tile being picked up/dropped
and some generic sliding action for the other tiles. The reordering
works so that all tiles retail their positions relative to each other
and simply fill up all the free space.