Don't create a layer with empty bounds
There is the case that a non-fbo layer with empty
bounds is created during createLayer() and so the
layer is cached and reused repetitively.

Since the purpose of the layer is to copy the
framebuffer into the layer, it is useless when
the bounds of the layer is empty.

This patch avoids such cases.

Change-Id: I363e86a9bc1650f9021942f8db75086cbde579f4
Signed-off-by: Dohyun Lee <dohyun.lee@lge.com>